requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
jinja2>=3.1.0
aiofiles
google-generativeai
openai>=1.0.0
chromadb
faiss-cpu
sentence-transformers
python-docx>=1.0.0
PyPDF2>=3.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0